rename package from ideal to idel. Add Ouath2 basic authentication workflow Simplify Basic Authentication config
