--- backend/src/main/kotlin/ideael/api/WebMvcConfig.kt
@@ -1,14 +1,0 @@
-package ideael.api
-
-import ideael.api.StringToIdeaSortingConverter
-import org.springframework.context.annotation.Configuration
-import org.springframework.format.FormatterRegistry
-import org.springframework.http.converter.HttpMessageConverter
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebMvcConfig : WebMvcConfigurer {
-    override fun addFormatters(registry: FormatterRegistry) {
-        registry.addConverter(StringToIdeaSortingConverter())
-    }
-}
