+++ backend/src/main/kotlin/idel/Application.kt
@@ -1,0 +1,13 @@
+package idel
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+
+@SpringBootApplication
+@ConfigurationPropertiesScan(basePackages = ["idel.infrastructure","idel.api"])
+class Application
+
+fun main() {
+    SpringApplication.run(Application::class.java)
+}
