+++ backend/src/main/kotlin/idel/api/WebMvcConfig.kt
@@ -1,0 +1,12 @@
+package idel.api
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.format.FormatterRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebMvcConfig : WebMvcConfigurer {
+    override fun addFormatters(registry: FormatterRegistry) {
+        registry.addConverter(StringToIdeaSortingConverter())
+    }
+}
