+++ backend/src/main/kotlin/idel/domain/User.kt
@@ -1,0 +1,11 @@
+package idel.domain
+
+typealias UserId = String
+
+
+interface User {
+    fun id() : String
+    val email : String
+    val profileName : String
+    val profilePhoto : String
+}
