+++ backend/src/main/kotlin/idel/infrastructure/repositories/CouchbaseProperties.kt
@@ -1,0 +1,14 @@
+package idel.infrastructure.repositories
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.ConstructorBinding
+
+
+@ConstructorBinding
+@ConfigurationProperties("couchbase")
+data class CouchbaseProperties(
+    val host : String,
+    val username : String,
+    val password : String,
+    val bucket : String ) {
+}
