+++ backend/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+ org.springframework.boot.env.PropertySourceLoader=com.github.zeldigas.spring.env.HoconPropertySourceLoader
