+++ backend/src/main/resources/application.conf
@@ -1,6 +1,6 @@
 # enalbe some features for testing, like as basic authentication. See testing.md for more details.
 # Don't use it in production, it's very dangerous!
-testmode=off
+testmode=on
 
 couchbase {
   host="localhost"
@@ -25,5 +25,12 @@ security {
       client-id=115577720302-jtm4mampbkbq62rerrb3thdcj5qq3egb.apps.googleusercontent.com
       client-secret=yLVkDuyUGMnM6J2iHtJ7t1a-
     }
+  }
+}
+
+spring.security.oauth2.client.registration {
+  google {
+    client-id=115577720302-jtm4mampbkbq62rerrb3thdcj5qq3egb.apps.googleusercontent.com
+    client-secret=yLVkDuyUGMnM6J2iHtJ7t1a-
   }
 }
