+++ docs/userinfo/facebook.json
@@ -1,0 +1,8 @@
+{
+"email": "leonid.vygovskiy@gmail.com",
+"name": "Leonid Vygovskiy",
+"first_name": "Leonid",
+"last_name": "Vygovskiy",
+"name_format": "{first} {last}",
+"id": "100001842993009"
+}
