+++ docs/userinfo/microsoft.json
@@ -1,0 +1,14 @@
+{
+  "@odata.context": "https://graph.microsoft.com/v1.0/$metadata#users/$entity",
+  "displayName": "Леонид Выговский",
+  "surname": "Выговский",
+  "givenName": "Леонид",
+  "id": "a450cda6820b7d71",
+  "userPrincipalName": "leonid.vygovskiy@gmail.com",
+  "businessPhones": [],
+  "jobTitle": null,
+  "mail": null,
+  "mobilePhone": null,
+  "officeLocation": null,
+  "preferredLanguage": null
+}
